requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
python-dotenv
azure-core
azure-ai-documentintelligence
langchain-openai
pandas
openpyxl
xlsxwriter
